pyproject.toml
[project]
name = "agentic-ai-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.5.5",
    "faiss-cpu>=1.12.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jupyterlab-widgets>=3.0.15",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langsmith>=0.4.27",
    "numpy>=2.3.2",
    "pydantic[email]>=2.11.9",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
]
